requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.3
google-cloud-sql-connector==1.2.2
pg8000==1.29.4  # For PostgreSQL, or use mysqlclient for MySQL